src/scripts/object-tracking-test/pyproject.toml
[project]
name = "object-tracking-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.4.3",
    "pyaml>=26.2.1",
    "pygame>=2.6.1",
]
